Trying different arguments for Backstop --skip-vr
The Pantheon template for these sites now includes a visual regression
step, provided via Backstop. This step can be skipped by using a skip-vr
flag in the commit message (as hopefully seen above).

The intent is that the review app is compared against the dev tier of
the pipeline, to guard against unintended visual changes.

The observed behavior, at the moment, is that Chrome is crashing on
CircleCI for unknown reasons. The usual search-the-error-message turns
up hints of memory problems in this scenario, so I tried adding various
flags to keep Chrome's resource use in check.

Those flags have not (yet) worked, so I'm squashing all those commits
and adding the --skip-vr flag to just blow past this step. I need to
get this merged and move onto other apps.